Swap output tensor indices in core.py
Originally, CalcTensor() function returned the temporal evolution of the  flows in arrays of shape (nt,N,N) where nt = number of time points and N = number of nodes.

This was a mistake. For coherents with C-style array format that is dominant in NumPy, now the output tensors are of shape (N,N,nt)