Remove 'DynCom' option
NetDynFlow is aimed to be an analysis tool for networks based on perturbation-induced flows. Therefore, the external input given to the network MUST be explicit and the user has to define it.
Function "DynCom()" was exactly the same as "DynFlow()" only that the input matrix Sigma defined as an identity matrix and hidden from the view to the user.
It is not a good idea we do this. User must always know which is the matrix of the perturbations.